// apps/frontend/packages/api/base/codegen/types/PlanCfoStatus2.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Cfo } from "./Cfo";
import type { PlanCfoStatusValue2 } from "./PlanCfoStatusValue2";
import type { UserSummary } from "./UserSummary";

export interface PlanCfoStatus2 {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type integer
    */
    planId: number;
    /**
     * @type integer
    */
    cfoId: number;
    /**
     * @type object
    */
    cfo: Cfo;
    /**
     * @type string
    */
    status: PlanCfoStatusValue2;
    /**
     * @type boolean
    */
    isRequired: boolean;
    /**
     * @type integer,null
    */
    decidedById?: number | null;
    decidedBy?: (UserSummary | null);
    /**
     * @type string,null, date-time
    */
    decidedAt?: string | null;
}