// apps/frontend/packages/api/base/codegen/types/PlanDetail.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Cfo } from "./Cfo";
import type { Filial } from "./Filial";
import type { Plan } from "./Plan";
import type { PlanCfoStatus2 } from "./PlanCfoStatus2";
import type { PlanDocumentSlot2 } from "./PlanDocumentSlot2";
import type { PlanHistoryEntry2 } from "./PlanHistoryEntry2";
import type { PlanRemark } from "./PlanRemark";
import type { PlanType } from "./PlanType";
import type { UserSummary } from "./UserSummary";

export type PlanDetail = (Plan & {
    /**
     * @type object
    */
    filial: Filial;
    /**
     * @type object
    */
    planType: PlanType;
    /**
     * @type object
    */
    author: UserSummary;
    /**
     * @type array
    */
    slots: PlanDocumentSlot2[];
    /**
     * @type array
    */
    cfoStatuses: PlanCfoStatus2[];
    /**
     * @type array
    */
    remarks: PlanRemark[];
    /**
     * @type array
    */
    history: PlanHistoryEntry2[];
    /**
     * @type boolean
    */
    packageComplete: boolean;
    /**
     * @type array
    */
    missingRequirements: string[];
    myCfoStatus?: (PlanCfoStatus2 | null);
    /**
     * @type integer | undefined
    */
    myOpenRemarksCount?: number;
    /**
     * @type boolean
    */
    isFilialOwner: boolean;
    /**
     * @type boolean
    */
    isCfoReviewer: boolean;
    /**
     * @type boolean
    */
    isDtoe: boolean;
    /**
     * @description ЦФО, доступные для отправки (linked_cfos), для формы «Направить»
     * @type array
    */
    availableCfos: Cfo[];
    /**
     * @description ЦФО в статусе RETURNED, для формы «Направить повторно»
     * @type array
    */
    returnedCfos: Cfo[];
});